Moved BackupKeysFragment to 'ui.activity.fragment' | #1253
